test(grok): split GrokProviderTests into per-class files
GrokProviderTests.swift had grown to ~600 lines, past the ~500-line
guideline, before the weekly-limit coverage lands. Pure move: the auth
store and log scanner suites (with their helpers) get their own files;
no test changes.
